--- .prettierignore
@@ -1,2 +1,0 @@
-dist
-node_modules
