--- .prettierrc
@@ -1,6 +1,0 @@
-{
-    "semi": false,
-    "trailingComma": "es5",
-    "singleQuote": true,
-    "useTabs": false
-}
