+++ bin/lsif-npm
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+
+require('../dist/src/npm/main.js')
