+++ bin/lsif-tsc
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+
+require('../dist/src/tsc/main.js')
