--- npm/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "references": [{ "path": "../shared" }],
-  "compilerOptions": {
-    "outDir": "dist"
-  },
-  "include": ["./**/*"]
-}
