+++ package.json
@@ -1,19 +1,22 @@
 {
   "$schema": "https://raw.githubusercontent.com/sourcegraph/sourcegraph/master/shared/src/schema/extension.schema.json",
   "name": "lsif-node",
-  "description": "",
+  "description": "A tool to create an LSIF dump for TypeScript projects.",
   "private": true,
   "sideEffects": false,
   "publisher": "sourcegraph",
-  "version": "0.0.0-DEVELOPMENT",
+  "version": "0.5.0",
   "license": "MIT",
   "repository": {
     "type": "git",
     "url": "https://github.com/sourcegraph/lsif-node"
   },
+  "bin": {
+    "lsif-tsc": "./bin/lsif-tsc",
+    "lsif-npm": "./bin/lsif-npm"
+  },
   "scripts": {
-    "prettier": "prettier --write --list-different '**/*.{ts,js?(on),md,yml}'",
-    "prettier-check": "yarn run prettier --write=false"
+    "build": "node ./node_modules/typescript/bin/tsc -b ."
   },
   "dependencies": {
     "lsif-protocol": "0.4.3",
@@ -24,18 +27,11 @@
     "vscode-uri": "^2.0.2"
   },
   "devDependencies": {
-    "@sourcegraph/prettierrc": "^3.0.2",
     "@sourcegraph/tsconfig": "^4.0.0",
     "@types/minimist": "^1.2.0",
     "@types/node": "13.7.4",
     "@types/npm": "^2.0.30",
-    "@types/shelljs": "^0.8.5",
     "@types/uuid": "^3.4.4",
-    "prettier": "^1.19.1",
-    "rimraf": "^2.6.3",
-    "shelljs": "^0.8.3",
-    "ts-node": "^8.6.2",
-    "typescript": "^3.7.5",
-    "yarn-deduplicate": "^1.1.1"
+    "typescript": "^3.7.5"
   }
 }
