+++ src/shared/consts.ts
@@ -1,0 +1,1 @@
+export const toolVersion = '0.5.0'
