+++ src/shared/is.ts
