+++ src/shared/linkedMap.ts
