+++ src/shared/paths.ts
