+++ src/shared/writer.ts
