+++ src/tsc/emitters/line.ts
@@ -2,19 +2,17 @@
  * Copyright (c) Microsoft Corporation. All rights reserved.
  * Licensed under the MIT License. See License.txt in the project root for license information.
  * ------------------------------------------------------------------------------------------ */
-import { Vertex, Edge, Id } from 'lsif-protocol';
-import { Writer } from '../../../shared/writer';
+import { Emitter, Create } from './emitter';
+import { Vertex, Edge } from 'lsif-protocol';
+import { Writer } from '../../shared/writer';
 
-export interface Emitter {
-	start(): void;
-	emit(element: Vertex | Edge): void;
-	end(): void;
-}
 
-export interface Create {
-	(writer: Writer, idGenerator: () => Id): Emitter;
-}
-
-export interface EmitterModule {
-	create: Create;
+export const create: Create = (writer: Writer): Emitter => {
+	return {
+		start: () => {},
+		emit: (element: Vertex | Edge) => {
+			writer.writeln(JSON.stringify(element, undefined, 0));
+		},
+		end: () => {}
+	};
 }
