+++ src/tsc/emitters/vis.ts
@@ -8,7 +8,7 @@ import { URI } from 'vscode-uri';
 
 import { Emitter, Create } from './emitter';
 import { Vertex, Edge, VertexLabels } from 'lsif-protocol';
-import { Writer } from '../../../shared/writer';
+import { Writer } from '../../shared/writer';
 
 interface VisNode {
 	id: number;
