+++ src/tsc/typescripts.ts
@@ -9,7 +9,7 @@ import * as crypto from 'crypto';
 
 import * as ts from 'typescript-lsif';
 
-import * as Is from '../../shared/is';
+import * as Is from '../shared/is';
 
 export type Declaration = ts.ModuleDeclaration | ts.ClassDeclaration | ts.InterfaceDeclaration | ts.TypeParameterDeclaration | ts.FunctionDeclaration | ts.MethodDeclaration |
 	ts.MethodSignature | ts.ParameterDeclaration;
