--- tsc/package.json
@@ -1,17 +1,0 @@
-{
-  "$schema": "https://raw.githubusercontent.com/sourcegraph/sourcegraph/master/shared/src/schema/extension.schema.json",
-  "name": "lsif-tsc",
-  "description": "A tool to create an LSIF dump for TypeScript projects.",
-  "sideEffects": false,
-  "publisher": "sourcegraph",
-  "version": "0.4.10",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/sourcegraph/lsif-node"
-  },
-  "main": "dist/main.js",
-  "scripts": {
-    "build": "node ../node_modules/typescript/bin/tsc -b ."
-  }
-}
