--- tsc/src/consts.ts
@@ -1,1 +1,0 @@
-export const toolVersion = require('../package.json').version;
