+++ tsconfig.json
@@ -1,14 +1,14 @@
 {
   "extends": "@sourcegraph/tsconfig",
-  "references": [{ "path": "./tsc" }, { "path": "./npm" }, { "path": "./shared" }],
   "compilerOptions": {
     "target": "es2019",
     "module": "commonjs",
     "allowJs": false,
     "moduleResolution": "node",
     "lib": ["esnext"],
-    "composite": true,
     "rootDir": ".",
-    "allowSyntheticDefaultImports": true
-  }
+    "allowSyntheticDefaultImports": true,
+    "outDir": "dist"
+  },
+  "include": ["./**/*"]
 }
